Creates a new root  class for converters dedicated to be used  in expression data mappings
